Clean Jupyter outputs and update README
Since output scrolling is not supported on Github, some of the Jupyter
outputs appeared too long for the page. These outputs have either been
omitted or shortened to make the notebook easier to read. Also Raw
NBconvert cells are not supported on Github so they have been converted
to code cells where the code is commented out so that it does not run.
Finally the README has been updated to explain a little about the
project.